pyproject.toml
[tool.poetry]
name = "appointment-scheduler"
version = "0.1.0"
description = ""
authors = ["Jose Roberto <joserobertomi@outlook.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.1.5"
asgiref = "^3.8.1"
gunicorn = "^23.0.0"
markdown = "^3.7"
pytz = "^2024.2"
django-cors-headers = "^4.6.0"
sqlparse = "^0.5.3"
psycopg2-binary = "^2.9.10"
djangorestframework = "^3.15.2"
pillow = "^11.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"